Fix uniqueColors
This should get all the keys out of the map, not all the pixels it has
seen in total.

+++ imgfiltrate.go
@@ -12,7 +12,7 @@ import (
 
 type Result struct {
 	ColorPercentage         float64 `json:"color_percentage"`
-	TotalColors             float64 `json:"total_colors"`
+	TotalColors             int     `json:"total_colors"`
 	AlphabeticCharacters    int     `json:"alphabetic_characters"`
 	NonAlphabeticCharacters int     `json:"non_alphabetic_characters"`
 	TotalCharacters         int     `json:"total_characters"`
